[refactor] 테스트 디렉토리 domain / service / controller 로 분리
